+++ blockfrost-client/src/Blockfrost/Client.hs
@@ -109,6 +109,12 @@ module Blockfrost.Client
   , getPoolBlocks'
   , getPoolUpdates
   , getPoolUpdates'
+    -- Cardano - Scripts
+  , listScripts
+  , listScripts'
+  , getScript
+  , getScriptRedeemers
+  , getScriptRedeemers'
     -- Cardano - Transactions
   , getTx
   , getTxUtxos
@@ -154,6 +160,7 @@ import Blockfrost.Client.Cardano.Ledger
 import Blockfrost.Client.Cardano.Metadata
 import Blockfrost.Client.Cardano.Network
 import Blockfrost.Client.Cardano.Pools
+import Blockfrost.Client.Cardano.Scripts
 import Blockfrost.Client.Cardano.Transactions
 import Blockfrost.Client.IPFS
 import Blockfrost.Client.NutLink